requirements/local.txt
-r base.txt  # include base requirements
alabaster==0.7.10
argh==0.26.2
astroid==1.5.3
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
docutils==0.14
idna==2.6
imagesize==0.7.1
isort==4.2.15
Jinja2==2.10
lazy-object-proxy==1.3.1
livereload==2.5.1
MarkupSafe==1.0
mccabe==0.6.1
pathtools==0.1.2
pkginfo==1.4.1
port-for==0.3.1
Pygments==2.2.0
pylint==1.7.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
PyYAML==3.12
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinx-autobuild==0.7.1
sphinxcontrib-websupport==1.0.1
tornado==4.5.2
tqdm==4.19.4
twine==1.9.1
urllib3==1.22
watchdog==0.8.3
wrapt==1.10.11
